move safe code out of mip
